requirements.txt
starlette
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv
sqlalchemy
alembic
databases
psycopg2-binary
asyncpg
aiosqlite
redis
celery
flower
httpx
orjson
fastapi-pagination
slowapi
pwdlib
pwdlib[argon2]
python-jose[cryptography]
python-multipart
email-validator
loguru
anyio
pytest
pytest-asyncio
tenacity
pydantic-settings
google-auth
google-auth-httplib2
requests
fastapi-mail
